chore (refs T22614): move store of procedure to client/js.
