+++ client/js/bundles/procedure/administrationSegmentsList.js
@@ -13,7 +13,7 @@
 import DpSlidebar from '@DpJs/components/core/DpSlidebar'
 import DpVersionHistory from '@DpJs/components/statement/statement/DpVersionHistory'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import SegmentFilterStore from '@DemosPlanProcedureBundle/store/SegmentFilter'
+import SegmentFilterStore from '@DpJs/store/procedure/SegmentFilter'
 import SegmentsList from '@DemosPlanProcedureBundle/components/SegmentsList/SegmentsList'
 
 const components = {
