+++ client/js/bundles/procedure/administrationStatementSegmentsList.js
@@ -12,8 +12,8 @@
  */
 
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import SegmentSlidebar from '@DemosPlanProcedureBundle/store/SegmentSlidebar'
-import SplitStatementStore from '@DemosPlanProcedureBundle/store/SplitStatementStore'
+import SegmentSlidebar from '@DpJs/store/procedure/SegmentSlidebar'
+import SplitStatementStore from '@DpJs/store/procedure/SplitStatementStore'
 import StatementSegmentsList from '@DemosPlanProcedureBundle/components/StatementSegmentsList/StatementSegmentsList'
 
 const components = {
