+++ client/js/store/procedure/SegmentSlidebar.js
